Added chat functionality

+++ IdentityService/IdentityService.Domain/User.cs
@@ -4,10 +4,10 @@ namespace IdentityService.Domain
 {
     public class User
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Email { get; set; }
         public string Name { get; set; }
-        public int Gender { get; set; } //TO-DO: consider enum?
+        public long Gender { get; set; } //TO-DO: consider enum?
         public DateTime? BirthDate { get; set; }
     }
 }
